% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{DRbayes}
\alias{DRbayes}
\title{Estimate causal effect using doubly robust estimators in high-dimensions}
\usage{
DRbayes(nScans = 20000, nBurn = 10000, thin = 10, y, x, t,
  whichCat = c(), y_type = "continuous", dfY = 1, dfT = 1, band = 3,
  thetaA = 1, thetaB = 0.2 * dim(x)[2])
}
\arguments{
\item{nScans}{The number of MCMC scans to run}

\item{nBurn}{The number of MCMC scans that will be dropped as a burn-in}

\item{thin}{This number represents how many iterations between each scan
that is kept}

\item{y}{The outcome to be analyzed}

\item{x}{An n by p matrix of potential confounders}

\item{t}{The treatment whose causal effect is to be estimated}

\item{whichCat}{A vector of indices that indicate which variables in x are categorical.
The default is c(), which means the program defaults to assuming all
covariates are continuous}

\item{y_type}{A categorical variable indicating whether y is binary or continuous.
Possible values are "binary" or "continuous", and the program defaults
to "continuous"}

\item{dfY}{the degrees of freedom of the splines used to model the relationship between
the covariates and the outcome. If the user wants to use gaussian process
priors instead of splines then dfY should be set to "GP"}

\item{dfT}{the degrees of freedom of the splines used to model the relationship between
the covariates and the treatment. If the user wants to use gaussian process
priors instead of splines then dfY should be set to "GP"}

\item{band}{The bandwidth parameter for the gaussian process kernel function}

\item{thetaA}{The first parameter of the beta prior on the overall sparsity level}

\item{thetaB}{The second parameter of the beta prior on the overall sparsity level}
}
\value{
A list of values that contain the treatment effect, confidence interval for the
        treatment effect, WAIC for the chosen treatment model and outcome model,
}
\description{
This function will take in the observed data and estimate a treatment effect.
y,x, and t must all be supplied, though all other parameters have pre-set values
the user can proceed with unless they wish to change the prior specification.
}
\examples{

## p can be larger than n, but we keep the number of covariates small here
## just for illustration so that the code will finish faster
n = 200
p = 20
x = matrix(rnorm(n*p), n, p)
t = rbinom(n, 1, p=pnorm(0.7*x[,1] + 0.3*x[,2]))
y = rnorm(n, mean=t + 0.3*x[,1] + 0.6*x[,2] + 0.5*x[,3], sd=1)

est = DRbayes(y=y, t=t, x=x, nScans=200, nBurn=100, thin=1)
}

